move sqlglot monkey patching to individual modules
